Improve bot logic

+++ astrogpt/models/advice.py
@@ -1,6 +1,6 @@
 from .base import Base
 
-from sqlalchemy import Integer, Text, DateTime, ForeignKey
+from sqlalchemy import Integer, Text, DateTime, ForeignKey, BigInteger
 from sqlalchemy.orm import Mapped, mapped_column, relationship
 from datetime import datetime
 
@@ -14,7 +14,7 @@ class Advice(Base):
     timestamp: Mapped[datetime] = mapped_column(DateTime, nullable=False, index=True)
 
     user_id: Mapped[int] = mapped_column(
-        Integer, ForeignKey("users.id"), nullable=False
+        BigInteger, ForeignKey("users.id"), nullable=False
     )
     user = relationship("User", back_populates="advices")
