+++ astrogpt/models/daily_forecast.py
@@ -1,6 +1,6 @@
 from .base import Base
 
-from sqlalchemy import Integer, Text, DateTime, ForeignKey
+from sqlalchemy import Integer, Text, DateTime, ForeignKey, BigInteger
 from sqlalchemy.orm import Mapped, mapped_column, relationship
 from datetime import datetime
 
@@ -13,7 +13,7 @@ class DailyForecast(Base):
     timestamp: Mapped[datetime] = mapped_column(DateTime, nullable=False, index=True)
 
     user_id: Mapped[int] = mapped_column(
-        Integer, ForeignKey("users.id"), nullable=False
+        BigInteger, ForeignKey("users.id"), nullable=False
     )
     user = relationship("User", back_populates="daily_forecasts")
